+++ src/functions/register_PointClouds.cpp
@@ -51,7 +51,7 @@ void pairAlign (
     const linkml::PointCloud::Cloud::Ptr cloud_tgt, 
     Eigen::Matrix4f &pairTransform, 
     std::optional<float> grid_size = {},
-    std::optional<uint8_t> confidence_filter = {}
+    std::optional<int32_t> confidence_filter = {}
     ){
     //
     // Downsample for consistency and speed
